refactor: code review
